refactor : 미팅 신청 상태 추가
